requirements.txt
requests
pandas
matplotlib 
python-dotenv
streamlit
plotly
streamlit-aggrid
